@@ -85,10 +85,11 @@
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<maven.build.timestamp.format>yyyyMMdd'T'HHmmss</maven.build.timestamp.format>
-		<maven.compiler.source>1.7</maven.compiler.source>
-		<maven.compiler.target>1.7</maven.compiler.target>
+		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>1.8</maven.compiler.target>
 		<hibernate-jpamodelgen.version>4.3.6.Final</hibernate-jpamodelgen.version>
-    <findbugs.skip>true</findbugs.skip>
+        <findbugs.skip>true</findbugs.skip>
+        <elasticsearch.version>2.1.2</elasticsearch.version>
 	</properties>
 
 	<dependencyManagement>
@@ -167,6 +168,7 @@
 		</pluginManagement>
 
 		<plugins>
+		 	<!--
 			<plugin>
 				<groupId>org.codehaus.mojo</groupId>
 				<artifactId>findbugs-maven-plugin</artifactId>
@@ -189,6 +191,7 @@
 					</execution>
 				</executions>
 			</plugin>
+			-->
 		</plugins>
 	</build>
 
