@@ -10,12 +10,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.joda.JodaModule;
+import com.fasterxml.jackson.datatype.jsr310.JSR310Module;
 
 import fi.otavanopisto.muikku.AbstractPyramusMocks;
 import fi.otavanopisto.pyramus.rest.model.ContactType;
@@ -100,7 +101,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   private static void mockStudyProgrammes(List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     StudyProgrammeCategory spc = new StudyProgrammeCategory(1l, "All Study Programmes", 1l, false);
     StudyProgramme sp = new StudyProgramme(1l, "test", "Test Study Programme", 1l, false);
@@ -134,7 +135,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
 
   public static void mockRoles(List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     CourseStaffMemberRole[] roles = { 
         new CourseStaffMemberRole(7l, "Opettaja"), 
@@ -177,7 +178,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
     emails.add("testuser@example.com");
     WhoAmI whoAmI = new WhoAmI((long) 1, "Test", "User", emails);
 
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     String whoAmIJson = objectMapper.writeValueAsString(whoAmI);
 
@@ -207,7 +208,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
     emails.add("teacher@example.com");
     WhoAmI whoAmI = new WhoAmI((long) 2, "Teacher", "User", emails);
 
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     String whoAmIJson = objectMapper.writeValueAsString(whoAmI);
 
@@ -237,7 +238,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
     emails.add("admin@example.com");
     WhoAmI whoAmI = new WhoAmI((long) 4, "Admin", "User", emails);
 
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     String whoAmIJson = objectMapper.writeValueAsString(whoAmI);
 
@@ -248,8 +249,8 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
         .withStatus(200)));
   }    
 
-  private static Person mockPerson(Long personId, DateTime birthday, String socialSecurityNumber, Sex sex, Long defaultUserId) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+  private static Person mockPerson(Long personId, OffsetDateTime birthday, String socialSecurityNumber, Sex sex, Long defaultUserId) throws JsonProcessingException {
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     Person person = new Person(personId, birthday, socialSecurityNumber, sex, false, "empty", defaultUserId);
     String personJson = objectMapper.writeValueAsString(person);
@@ -264,7 +265,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   private static StaffMember mockStaffMember(Long personId, Long staffMemberId, String firstName, String lastName, String email, UserRole role, List<String> tags, Map<String, String> variables, List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     StaffMember staffMember = new StaffMember(staffMemberId, personId, null, firstName, lastName, null, 
         role, tags, variables);
@@ -301,8 +302,8 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
     return staffMember;
   }
   
-  public static Student mockStudent(Long personId, Long studentId, String firstName, String lastName, String email, List<String> tags, Map<String, String> variables, List<String> payloads, DateTime studyStartDate, DateTime studyEndDate) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+  public static Student mockStudent(Long personId, Long studentId, String firstName, String lastName, String email, List<String> tags, Map<String, String> variables, List<String> payloads, OffsetDateTime studyStartDate, OffsetDateTime studyEndDate) throws JsonProcessingException {
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     Student student = new Student(
         studentId,
@@ -365,9 +366,9 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   public static void mockUsers(List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
-    DateTime birthday = new DateTime(1990, 2, 2, 0, 0, 0, 0);
+    OffsetDateTime birthday = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
 
     Person person = mockPerson(1l, birthday, "345345-3453", Sex.MALE, 1l);
     Person staff1 = mockPerson(2l, birthday, "345345-3453", Sex.MALE, 2l);
@@ -443,13 +444,13 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
         .withStatus(200)));
   }
   
-  private static DateTime getNextYear() {
-    DateTime result = new DateTime();
+  private static OffsetDateTime getNextYear() {
+    OffsetDateTime result = OffsetDateTime.now(ZoneOffset.UTC);
     return result.plusYears(1);
   }
   
-  private static DateTime toDate(int year, int month, int day) {
-    return new DateTime(year, month, day, 0, 0, 0, 0);
+  private static OffsetDateTime toDate(int year, int month, int day) {
+    return OffsetDateTime.of(year, month, day, 0, 0, 0, 0, ZoneOffset.UTC);
   }
 
   public static void mockWorkspaces(List<String> payloads) throws JsonProcessingException {
@@ -457,9 +458,9 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   public static void mockWorkspace(Long id, List<String> payloads) throws JsonProcessingException {
-    DateTime created = new DateTime(1990, 2, 2, 0, 0, 0, 0);
-    DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
-    DateTime end = new DateTime(2050, 1, 1, 0, 0, 0, 0);
+    OffsetDateTime created = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime begin = OffsetDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime end = OffsetDateTime.of(2050, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
 
     Course course = new Course(
         id,
@@ -492,7 +493,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
        null,
        null);
   
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     String courseJson = objectMapper.writeValueAsString(course);
     
     stubFor(get(urlEqualTo(String.format("/1/courses/courses/%d", id)))
@@ -537,7 +538,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
 
   public static void mockCommons() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     EducationType educationType = new EducationType((long) 1, "testEduType", "ET", false);
     String educationTypeJson = objectMapper.writeValueAsString(educationType);
@@ -595,9 +596,9 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   public static void mockUserGroups(List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
-    DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
+    OffsetDateTime begin = OffsetDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
     Long creatorId = 1l;
     Long groupId = 2l;
     
@@ -656,7 +657,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   public static void mockWorkspaceUsers(List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     Long courseId = 1l;
     Long teacherRoleId = 7l;
@@ -678,7 +679,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
 
     addPayload(payloads, objectMapper.writeValueAsString(new WebhookCourseStaffMemberCreatePayload(1l, courseId, staffMember.getId())));
     
-    DateTime enrolmentTime = new DateTime(1999, 1, 1, 0, 0, 0, 0);
+    OffsetDateTime enrolmentTime = OffsetDateTime.of(1999, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
     CourseStudent student = new CourseStudent(2l, courseId, 1l, enrolmentTime, false, null, null, false, null, null);
     CourseStudent[] students = { student };
 
@@ -698,7 +699,7 @@ public class PyramusMocksRest extends AbstractPyramusMocks {
   }
   
   private static void addPayload(List<String> payloads, Object obj) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     addPayload(payloads, objectMapper.writeValueAsString(obj));
   }
