@@ -11,12 +11,13 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.joda.JodaModule;
+import com.fasterxml.jackson.datatype.jsr310.JSR310Module;
 
 import fi.otavanopisto.muikku.AbstractPyramusMocks;
 import fi.otavanopisto.muikku.TestUtilities;
@@ -70,7 +71,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   private static void loginMock(Long userId, String email, String firstName, String lastName) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     stubFor(get(urlEqualTo("/dnm")).willReturn(aResponse().withHeader("Content-Type", "application/json").withBody("").withStatus(204)));
 
@@ -106,12 +107,12 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void personsPyramusMocks() throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     Map<String, String> variables = null;
     List<String> tags = null;
     
-    DateTime birthday = new DateTime(1990, 2, 2, 0, 0, 0, 0);
+    OffsetDateTime birthday = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
     Person person = mockPerson(1l, birthday, "030545-3453", fi.otavanopisto.pyramus.rest.model.Sex.MALE, 1l);
     
     Student student = new Student((long) 1, (long) 1, "Test", "User", null, null, null, null, null, null, null, null,
@@ -146,7 +147,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
         .withStatus(200)));
     
     /* Student #2 for workspace #2*/
-    DateTime birthday2 = new DateTime(1992, 2, 2, 0, 0, 0, 0);
+    OffsetDateTime birthday2 = OffsetDateTime.of(1992, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
     Person person2 = mockPerson(5l, birthday2, "021092-2112", fi.otavanopisto.pyramus.rest.model.Sex.MALE, 5l);
     
     Student student2 = new Student((long) 5, (long) 5, "Second", "User", null, null, null, null, null, null, null, null,
@@ -386,7 +387,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
     mockContactTypes();
     mockStudyProgrammes();
      
-    CourseStudent courseStudent = new CourseStudent(3l, 1l, 1l, new DateTime(2010, 2, 2, 0, 0, 0, 0), false, null, null, null, null, null);
+    CourseStudent courseStudent = new CourseStudent(3l, 1l, 1l, OffsetDateTime.of(2010, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC), false, null, null, null, null, null);
     CourseStudent[] csArray = { courseStudent };
     String coursestudentArrayJson = objectMapper.writeValueAsString(csArray);
     String courseStudenJson = objectMapper.writeValueAsString(courseStudent);
@@ -409,7 +410,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
         .withBody(courseStudenJson)
         .withStatus(200)));
     
-    DateTime assessmentCreated = new DateTime(2015, 2, 2, 0, 0, 0, 0);
+    OffsetDateTime assessmentCreated = OffsetDateTime.of(2015, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
     CourseAssessment courseAssessment = new CourseAssessment(1l, courseStudent.getId(), 1l, 1l, 4l, assessmentCreated, "Test evaluation.");
     
     stubFor(post(urlMatching(String.format("/1/students/students/%d/courses/%d/assessments/", student.getId(), courseStudent.getCourseId())))
@@ -424,7 +425,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
         .withBody(objectMapper.writeValueAsString(new ArrayList<>()))
         .withStatus(200)));
     
-    CourseStudent courseStudent2 = new CourseStudent(2l, 2l, 5l, new DateTime(2010, 2, 2, 0, 0, 0, 0), false, null, null, null, null, null);
+    CourseStudent courseStudent2 = new CourseStudent(2l, 2l, 5l, OffsetDateTime.of(2010, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC), false, null, null, null, null, null);
     CourseStudent[] csArray2 = { courseStudent2 };
     String coursestudentArrayJson2 = objectMapper.writeValueAsString(csArray2);
     String courseStudenJson2 = objectMapper.writeValueAsString(courseStudent2);
@@ -496,7 +497,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   
   @SuppressWarnings("unused")
   private static StaffMember mockStaffMember(Long personId, Long staffMemberId, String firstName, String lastName, String email, UserRole role, List<String> tags, Map<String, String> variables, List<String> payloads) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     StaffMember staffMember = new StaffMember(staffMemberId, personId, null, firstName, lastName, null, 
         role, tags, variables);
@@ -534,7 +535,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   private static void mockStudyProgrammes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     StudyProgrammeCategory spc = new StudyProgrammeCategory(1l, "All Study Programmes", 1l, false);
     StudyProgramme sp = new StudyProgramme(1l, "test", "Test Study Programme", 1l, false);
@@ -572,10 +573,10 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void workspacePyramusMock(Long id, String name, String description) throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-    DateTime created = new DateTime(1990, 2, 2, 0, 0, 0, 0);
-    DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
-    DateTime end = new DateTime(2050, 1, 1, 0, 0, 0, 0);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    OffsetDateTime created = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime begin = OffsetDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime end = OffsetDateTime.of(2050, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
 
     Course course = new Course(id, name, created, created, description, false, 1, 
       (long) 25, begin, end, "test extension", (double) 15, (double) 45, (double) 45,
@@ -620,10 +621,10 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
 
   public static void studentGroupsMocks() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-    DateTime created = new DateTime(1990, 2, 2, 0, 0, 0, 0);
-    DateTime begin = new DateTime(2000, 1, 1, 0, 0, 0, 0);
-    DateTime lastmodified = new DateTime(2000, 1, 1, 0, 0, 0, 0);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    OffsetDateTime created = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime begin = OffsetDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
+    OffsetDateTime lastmodified = OffsetDateTime.of(2000, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC);
     StudentGroup studentGroupStudents = new StudentGroup((long) 1, "Opiskelijat", "Spring 2015 Students", begin, (long) 1, created, (long) 1, lastmodified, null, false);
     StudentGroup studentGroupAnother = new StudentGroup((long) 2, "Opiskelijat 2", "Spring 2015 Students 2", begin, (long) 1, created, (long) 1, lastmodified, null, false);
     StudentGroup[] studentGroupArray = {studentGroupStudents, studentGroupAnother};
@@ -636,12 +637,12 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void adminMock() throws Exception {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);    
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);    
     
     Map<String, String> variables = null;
     List<String> tags = null;
     
-    DateTime birthday = new DateTime(1990, 2, 2, 0, 0, 0, 0);
+    OffsetDateTime birthday = OffsetDateTime.of(1990, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
     
     Person staff3 = mockPerson((long) 4, birthday, "345345-3453", fi.otavanopisto.pyramus.rest.model.Sex.MALE, (long) 4);
     
@@ -739,8 +740,8 @@ public class PyramusMocks extends AbstractPyramusMocks {
  
   }
   
-  private static Person mockPerson(Long personId, DateTime birthday, String socialSecurityNumber, Sex sex, Long defaultUserId) throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+  private static Person mockPerson(Long personId, OffsetDateTime birthday, String socialSecurityNumber, Sex sex, Long defaultUserId) throws JsonProcessingException {
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     Person person = new Person(personId, birthday, socialSecurityNumber, sex, false, "empty", defaultUserId);
     String personJson = objectMapper.writeValueAsString(person);
@@ -755,7 +756,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void mockCommons() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
 
     Subject subject = new Subject((long) 1, "tc_11", "Test course", (long) 1, false);
     String subjectJson = objectMapper.writeValueAsString(subject);
@@ -856,7 +857,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
 
   public static void mockCourseTypes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     fi.otavanopisto.pyramus.rest.model.CourseType courseType = new fi.otavanopisto.pyramus.rest.model.CourseType((long) 1, "Nonstop", false);
     CourseType[] courseTypeArray = { courseType };
@@ -879,7 +880,7 @@ public class PyramusMocks extends AbstractPyramusMocks {
   }
   
   public static void mockContactTypes() throws JsonProcessingException {
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
     
     ContactType contactType = new ContactType((long)1, "Koti", false, false);
     ContactType[] contactTypes = { contactType };
@@ -901,9 +902,9 @@ public class PyramusMocks extends AbstractPyramusMocks {
   public static void mockAssessedStudent1Workspace1(MockCourseStudent courseStudent, Long assessorId) throws JsonProcessingException {
     // TODO: Move to new mocker.
     
-    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JodaModule()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
-//    CourseStudent courseStudent = new CourseStudent(3l, 1l, 1l, new DateTime(2010, 2, 2, 0, 0, 0, 0), false, null, null, null, null, null);
-    DateTime assessmentCreated = new DateTime(2015, 2, 2, 0, 0, 0, 0);
+    ObjectMapper objectMapper = new ObjectMapper().registerModule(new JSR310Module()).disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+//    CourseStudent courseStudent = new CourseStudent(3l, 1l, 1l, OffsetDateTime.of(2010, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC), false, null, null, null, null, null);
+    OffsetDateTime assessmentCreated = OffsetDateTime.of(2015, 2, 2, 0, 0, 0, 0, ZoneOffset.UTC);
     CourseAssessment courseAssessment = new CourseAssessment(1l, courseStudent.getId(), 1l, 1l, assessorId, assessmentCreated, "Test evaluation.");
     List<CourseAssessment> courseAssessments = new ArrayList<CourseAssessment>();
     courseAssessments.add(courseAssessment);
