@@ -1,6 +1,7 @@
 package fi.otavanopisto.muikku.plugins.schooldatalocal.entities;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.plugins.schooldatalocal.LocalUserSchoolDataController;
 import fi.otavanopisto.muikku.schooldata.entity.AbstractUser;
@@ -10,7 +11,7 @@ public class LocalUserImpl  extends AbstractUser implements User {
 
   public LocalUserImpl(String identifier, String firstName, String lastName, String displayName,
       String studyProgrammeName, String nationality, String language, String municipality, String school,
-      DateTime studyStartDate, DateTime studyEndDate, DateTime studyTimeEnd, String curriculumIdentifier) {
+      OffsetDateTime studyStartDate, OffsetDateTime studyEndDate, OffsetDateTime studyTimeEnd, String curriculumIdentifier) {
     super(identifier, firstName, lastName, displayName, studyProgrammeName, nationality, language, municipality, school,
         studyStartDate, studyEndDate, studyTimeEnd, false, false, false, false, false, curriculumIdentifier);
   }
