@@ -1,13 +1,14 @@
 package fi.otavanopisto.muikku.schooldata.entity;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 
 public abstract class AbstractWorkspaceUser implements WorkspaceUser {
 
   public AbstractWorkspaceUser(SchoolDataIdentifier identifier, SchoolDataIdentifier userIdentifier, SchoolDataIdentifier workspaceIdentifier,
-      SchoolDataIdentifier roleIdentifier, DateTime enrolmentTime) {
+      SchoolDataIdentifier roleIdentifier, OffsetDateTime enrolmentTime) {
     super();
     this.identifier = identifier;
     this.userIdentifier = userIdentifier;
@@ -37,7 +38,7 @@ public abstract class AbstractWorkspaceUser implements WorkspaceUser {
   }
   
   @Override
-  public DateTime getEnrolmentTime() {
+  public OffsetDateTime getEnrolmentTime() {
     return enrolmentTime;
   }
   
@@ -45,5 +46,5 @@ public abstract class AbstractWorkspaceUser implements WorkspaceUser {
   private SchoolDataIdentifier userIdentifier;
   private SchoolDataIdentifier workspaceIdentifier;
   private SchoolDataIdentifier roleIdentifier;
-  private DateTime enrolmentTime;
+  private OffsetDateTime enrolmentTime;
 }
