@@ -1,6 +1,7 @@
 package fi.otavanopisto.muikku.schooldata.entity;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
 
@@ -10,6 +11,6 @@ public interface WorkspaceUser extends SchoolDataEntity {
 	public SchoolDataIdentifier getUserIdentifier();
 	public SchoolDataIdentifier getWorkspaceIdentifier();
 	public SchoolDataIdentifier getRoleIdentifier();
-	public DateTime getEnrolmentTime();
+	public OffsetDateTime getEnrolmentTime();
 	
 }
