@@ -4,7 +4,8 @@ import java.util.List;
 
 import javax.inject.Inject;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.calendar.CalendarEvent;
 import fi.otavanopisto.muikku.calendar.CalendarEventAttendee;
@@ -79,7 +80,7 @@ public class GoogleCalendarServiceProvider implements CalendarServiceProvider {
   }
 
   @Override
-  public List<CalendarEvent> listEvents(DateTime minTime, DateTime maxTime, String... calendarId) throws CalendarServiceException {
+  public List<CalendarEvent> listEvents(OffsetDateTime minTime, OffsetDateTime maxTime, String... calendarId) throws CalendarServiceException {
     return calendarClient.listEvents(minTime, maxTime, calendarId);
   }
 
