@@ -1,6 +1,7 @@
 package fi.otavanopisto.muikku.plugins.schooldatapyramus.entities;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.plugins.schooldatapyramus.SchoolDataPyramusPluginDescriptor;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
@@ -11,7 +12,7 @@ public class PyramusTransferCredit extends AbstractTransferCredit {
   public PyramusTransferCredit() {
   }
   
-  public PyramusTransferCredit(SchoolDataIdentifier identifier, SchoolDataIdentifier studentIdentifier, DateTime date,
+  public PyramusTransferCredit(SchoolDataIdentifier identifier, SchoolDataIdentifier studentIdentifier, OffsetDateTime date,
       SchoolDataIdentifier gradeIdentifier, SchoolDataIdentifier gradingScaleIdentifier, String verbalAssessment, SchoolDataIdentifier assessorIdentifier,
       String courseName, Integer courseNumber, Double length, SchoolDataIdentifier lengthUnitIdentifier,
       SchoolDataIdentifier schoolIdentifier, SchoolDataIdentifier subjectIdentifier) {
