@@ -2,7 +2,8 @@ package fi.otavanopisto.muikku.plugins.schooldatapyramus.entities;
 
 import java.util.Date;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.plugins.schooldatapyramus.SchoolDataPyramusPluginDescriptor;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
@@ -17,7 +18,7 @@ public class PyramusWorkspace extends AbstractWorkspace {
   public PyramusWorkspace(String identifier, String name, String nameExtension, String viewLink,
       SchoolDataIdentifier workspaceTypeId, String courseIdentifierIdentifier, String description,
       String subjectIdentifier, SchoolDataIdentifier educationTypeIdentifier, Date modified, Double length,
-      String lengthUnitIdentifier, DateTime beginDate, DateTime endDate, boolean archived,
+      String lengthUnitIdentifier, OffsetDateTime beginDate, OffsetDateTime endDate, boolean archived,
       boolean evaluationFeeApplicable, SchoolDataIdentifier curriculumIdentifier) {
     super(identifier, name, nameExtension, viewLink, workspaceTypeId, courseIdentifierIdentifier, description,
         subjectIdentifier, educationTypeIdentifier, modified, length, lengthUnitIdentifier, beginDate, endDate, archived,
