@@ -1,6 +1,7 @@
 package fi.otavanopisto.muikku.plugins.schooldatapyramus.entities;
 
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.plugins.schooldatapyramus.SchoolDataPyramusPluginDescriptor;
 import fi.otavanopisto.muikku.schooldata.SchoolDataIdentifier;
@@ -10,7 +11,7 @@ import fi.otavanopisto.muikku.schooldata.entity.WorkspaceUser;
 public class PyramusWorkspaceUser extends AbstractWorkspaceUser implements WorkspaceUser {
 
 	public PyramusWorkspaceUser(SchoolDataIdentifier identifier, SchoolDataIdentifier userIdentifier, SchoolDataIdentifier workspaceIdentifier,
-      SchoolDataIdentifier roleIdentifier, DateTime enrolmentTime) {
+      SchoolDataIdentifier roleIdentifier, OffsetDateTime enrolmentTime) {
     super(identifier, userIdentifier, workspaceIdentifier, roleIdentifier, enrolmentTime);
   }
 
