@@ -19,7 +19,8 @@ import javax.inject.Inject;
 import org.apache.commons.lang3.LocaleUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.math.NumberUtils;
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.controller.PluginSettingsController;
 import fi.otavanopisto.muikku.i18n.LocaleController;
@@ -135,7 +136,7 @@ public class RequestedAssessmentSupplementationsNotificationStrategy extends Abs
           if (workspaceAssessments != null && !workspaceAssessments.isEmpty()) {
             WorkspaceAssessment assessment = workspaceAssessments.get(0); //TODO: loop and find latest
             Date assessmentDate = assessment.getDate();
-            if (assessmentDate != null && assessmentDate.before(new DateTime().minusDays(NOTIFICATION_THRESHOLD_DAYS).toDate())) {
+            if (assessmentDate != null && assessmentDate.before(Date.from(OffsetDateTime.now().minusDays(NOTIFICATION_THRESHOLD_DAYS).toInstant()))) {
               GradingScale gradingScale = gradingController.findGradingScale(assessment.getGradingScaleSchoolDataSource(), assessment.getGradingScaleIdentifier());
               GradingScaleItem grade = gradingController.findGradingScaleItem(gradingScale, assessment.getGradeSchoolDataSource(), assessment.getGradeIdentifier());
               
