@@ -19,7 +19,8 @@ import javax.inject.Inject;
 import org.apache.commons.lang3.LocaleUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.math.NumberUtils;
-import org.joda.time.DateTime;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
 
 import fi.otavanopisto.muikku.controller.PluginSettingsController;
 import fi.otavanopisto.muikku.i18n.LocaleController;
@@ -79,8 +80,8 @@ public class StudyTimeNotificationStrategy extends AbstractTimedNotificationStra
       return;
     }
     
-    Date studyTimeEnds = new DateTime().plusDays(NOTIFICATION_THRESHOLD_DAYS_LEFT).toDate();
-    List<SchoolDataIdentifier> studentIdentifierAlreadyNotified = studyTimeLeftNotificationController.listNotifiedSchoolDataIdentifiersAfter(new DateTime().minusDays(NOTIFICATION_THRESHOLD_DAYS_LEFT).toDate());
+    Date studyTimeEnds = Date.from(OffsetDateTime.now().plusDays(NOTIFICATION_THRESHOLD_DAYS_LEFT).toInstant());
+    List<SchoolDataIdentifier> studentIdentifierAlreadyNotified = studyTimeLeftNotificationController.listNotifiedSchoolDataIdentifiersAfter(Date.from(OffsetDateTime.now().minusDays(NOTIFICATION_THRESHOLD_DAYS_LEFT).toInstant()));
     SearchResult searchResult = studyTimeLeftNotificationController.searchActiveStudentIds(groups, FIRST_RESULT + offset, MAX_RESULTS, studentIdentifierAlreadyNotified, studyTimeEnds);
     
     if (searchResult.getFirstResult() + MAX_RESULTS >= searchResult.getTotalHitCount()) {
