@@ -57,7 +57,7 @@ public class UserRESTServiceTestsIT extends AbstractRESTTest {
     String expected = "{'id':1,'firstName':'Test','lastName':'User','hasImage':false}";
     JSONAssert.assertEquals(expected, response.body().asString(), STRICT_JSON);
   }
-
+  
   @Test
   public void testSearchUsers() throws NoSuchFieldException {
     Response response = asAdmin()
@@ -178,5 +178,4 @@ public class UserRESTServiceTestsIT extends AbstractRESTTest {
     String expected = "[{'id':1,'firstName':'Test','lastName':'User','hasImage':false,'nationality':null,'language':null,'municipality':null,'school':null,'email':'te...@example.com'}]";
     JSONAssert.assertEquals(expected, response.body().asString(), STRICT_JSON);
   }
-
 }
